Documentation (#40)
* init readme update

* fix newlines

* cleanup readme

* remove unnecessary text "function" "abstract method"

* comment in sim, astropy, matplob

* delete unnecessary words in docstrings, add docstrings to states_to_df and df_to_csv

* move docstring in line of triple quotes

* Put docstrings onto new line

* remove unnecessary words

* Remove unnecessary words

* move docstrings to a new line, delete unnecessary args and returns section

* remove types from matplotlib util - weren't too helpful anyway

* fix pyright fr this time

* clarify readme

* fix clown mistake

* change json example

* Revert "merge"

This reverts commit 36ec4e5e508f7b3296bee17b9aebb4d1821baa1e, reversing
changes made to 520de3022d939d1206e48c031bb54773d7d37375.

* init readme update

* fix newlines

* cleanup readme

* remove unnecessary text "function" "abstract method"

* comment in sim, astropy, matplob

* delete unnecessary words in docstrings, add docstrings to states_to_df and df_to_csv

* move docstring in line of triple quotes

* Put docstrings onto new line

* remove unnecessary words

* remove types from matplotlib util - weren't too helpful anyway

* Remove unnecessary words

* move docstrings to a new line, delete unnecessary args and returns section

* fix pyright fr this time

* clarify readme

* fix clown mistake

* change json example

* Revert "merge"

This reverts commit 36ec4e5e508f7b3296bee17b9aebb4d1821baa1e, reversing
changes made to 520de3022d939d1206e48c031bb54773d7d37375.
