--- configs/attitude_dynamics_test.json
@@ -1,13 +1,0 @@
-{
-    "initial_condition": {
-        "x": 100000000.0,
-        "ang_vel_x": 4.5,
-        "ang_vel_y": 1.0,
-        "time": 0.0,
-        "quat_v1": 1.0,
-        "quat_v2": 0.0,
-        "quat_v3": 0.0,
-        "quat_r" :0.0
-    },
-    "models" : ["att", "gyro"]
- }
