+++ configs/freefall.json
@@ -1,15 +1,13 @@
 {
+    "description": "This config models the satellite in free fall motion. The sim runs until the satellite hits the earth. The visualization should display a relatively straight path towards the earth. ",
+	
     "initial_condition": {
 
         "x": 100000000.0,
         "y": 1000.0,
         "z": 1000.0,
         "ang_vel_x": 4.5,
-        "time": 0.0,
-        "quat_v1": 1.0,
-        "quat_v2": 0.0,
-        "quat_v3": 0.0,
-        "quat_r" :0.0
+        "time": 0.0
     },
-    "models" : ["pos", "att", "gyro"]
+    "models" : ["pos", "gyro"]
  }
