+++ configs/test_angles.json
@@ -5,10 +5,10 @@
         "ang_vel_y" : 5.0, 
         "ang_vel_z" : 5.0, 
         
-        "quat_v1" : 10.0, 
-        "quat_v2" : 10.0, 
-        "quat_v3" : 10.0, 
-        "quat_r" : 10.0, 
+        "gnc_pos_q1" : 10.0, 
+        "gnc_pos_q2" : 10.0, 
+        "gnc_pos_q3" : 10.0, 
+        "gnc_pos_q4" : 10.0, 
 
         "vel_x" : 3.0, 
         "vel_y" : 3.0,
