+++ src/core/models/model.py
@@ -1,9 +1,38 @@
-from abc import abstractmethod
+from abc import ABC, abstractmethod
 from typing import Dict, Any, Type, Union
-from core.state.statetime import StateTime
+from core.state import StateTime
 from core.parameters import Parameters
 from utils.constants import State_Type
-from core.models.model_base import Model
+
+
+class Model(ABC):
+    """
+    Abstract Base class for all models this sim uses.
+    """
+
+    def __init__(self, parameters: Parameters) -> None:
+        """Model __init__
+        All models will be dependent on some parameters, so we load them in
+        here.
+        Args:
+            parameters (Parameters): Instance of the parameters class gets
+            passed in to be accessible by the model.
+        """
+        self._parameters = parameters
+
+    @abstractmethod
+    def evaluate(self, state_time: StateTime):
+        """
+        Evaluates the model based on the
+            current state.
+        An instance of State is required to evaluate the model, (because each
+            model should be dependent on the state of the system.)
+        Args:
+            state (State): a instance of a State class.
+
+        Returns:
+            _type_: Defined in concrete instantiation of subclasses.
+        """
 
 
 class EnvironmentModel(Model):
@@ -15,7 +44,7 @@ class EnvironmentModel(Model):
 
     @abstractmethod
     def d_state(self, state_time: StateTime) -> Dict[str, State_Type]:
-        """Function which evaluates the differential equation:
+        """Evaluates the differential equation:
             dy / dt = f(t, y)
             for the current state. "y" is a state vector (not just one variable)
