+++ tests/statetime_performance.py
@@ -1,6 +1,5 @@
 import datetime
-from core.state.state import State
-from core.state.statetime import StateTime
+from core.state import State, StateTime
 
 N_TESTS = 100000
